Fix clang-tidy warnings for clang 7
Disable clang-analyzer because we run analyzer as seperate job.
Disable cppcoreguidelines-avoid-goto because it is easy to spot
in code-review and has very rare good uses.

+++ src/gromacs/math/functions.h
@@ -474,22 +474,22 @@ constexpr int64_t exactDiv(int64_t a, int64_t b)
  */
 static inline int roundToInt(float x)
 {
-    return rintf(x);
+    return static_cast<int>(rintf(x));
 }
 //! Round double to int
 static inline int roundToInt(double x)
 {
-    return rint(x);
+    return static_cast<int>(rint(x));
 }
 //! Round float to int64_t
 static inline int64_t roundToInt64(float x)
 {
-    return rintf(x);
+    return static_cast<int>(rintf(x));
 }
 //! Round double to int64_t
 static inline int64_t roundToInt64(double x)
 {
-    return rint(x);
+    return static_cast<int>(rint(x));
 }
 
 } // namespace gmx
